daemon_util/requirements.txt
envoy==0.0.3
grpcio==1.20.1
grpcio-tools==1.20.1
libvirt-python==5.3.0
lxml==4.3.3
protobuf==3.7.1
psutil==5.6.2
six==1.12.0
